requirements.txt
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
awscrt==0.20.9
awsiotsdk==1.21.4
build==0.10.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dill==0.3.7
dnspython==2.4.2
-e git+https://github.com/CharlyMercury/ed_websockets.git@9dc897a75f910af06eeb30745378eeb9b69dc135#egg=Ed_WebSockets
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
orjson==3.9.5
packaging==23.1
platformdirs==3.10.0
pydantic==2.2.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.6.1
pylint==2.17.5
pyproject_hooks==1.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
